Restructured source interface
